fixes + catch-all for ownables
