+++ typescript/infra/src/utils/safe.ts
@@ -7,6 +7,7 @@ import {
 } from '@safe-global/safe-core-sdk-types';
 import chalk from 'chalk';
 import { BigNumber, ethers } from 'ethers';
+import { formatUnits } from 'ethers/lib/utils.js';
 
 import {
   ChainNameOrId,
@@ -14,9 +15,16 @@ import {
   getSafe,
   getSafeService,
 } from '@hyperlane-xyz/sdk';
-import { Address, CallData, eqAddress, retryAsync } from '@hyperlane-xyz/utils';
+import {
+  Address,
+  CallData,
+  eqAddress,
+  retryAsync,
+  rootLogger,
+} from '@hyperlane-xyz/utils';
 
 import safeSigners from '../../config/environments/mainnet3/safe/safeSigners.json' assert { type: 'json' };
+// eslint-disable-next-line import/no-cycle
 import { AnnotatedCallData } from '../govern/HyperlaneAppGovernor.js';
 
 export async function getSafeAndService(
@@ -82,7 +90,7 @@ export async function executeTx(
 
   await safeSdk.executeTransaction(safeTransaction);
 
-  console.log(
+  rootLogger.info(
     chalk.green.bold(`Executed transaction ${safeTxHash} on ${chain}`),
   );
 }
@@ -122,7 +130,7 @@ export async function proposeSafeTransaction(
     senderSignature: senderSignature.data,
   });
 
-  console.log(
+  rootLogger.info(
     chalk.green(`Proposed transaction on ${chain} with hash ${safeTxHash}`),
   );
 }
@@ -143,7 +151,7 @@ export async function deleteAllPendingSafeTxs(
   });
 
   if (!pendingTxsResponse.ok) {
-    console.error(
+    rootLogger.error(
       chalk.red(`Failed to fetch pending transactions for ${safeAddress}`),
     );
     return;
@@ -156,7 +164,7 @@ export async function deleteAllPendingSafeTxs(
     await deleteSafeTx(chain, multiProvider, safeAddress, tx.safeTxHash);
   }
 
-  console.log(
+  rootLogger.info(
     `Deleted all pending transactions on ${chain} for ${safeAddress}\n`,
   );
 }
@@ -177,7 +185,7 @@ export async function getSafeTx(
   });
 
   if (!txDetailsResponse.ok) {
-    console.error(
+    rootLogger.error(
       chalk.red(`Failed to fetch transaction details for ${safeTxHash}`),
     );
     return;
@@ -205,7 +213,7 @@ export async function deleteSafeTx(
   });
 
   if (!txDetailsResponse.ok) {
-    console.error(
+    rootLogger.error(
       chalk.red(`Failed to fetch transaction details for ${safeTxHash}`),
     );
     return;
@@ -215,21 +223,23 @@ export async function deleteSafeTx(
   const proposer = txDetails.proposer;
 
   if (!proposer) {
-    console.error(chalk.red(`No proposer found for transaction ${safeTxHash}`));
+    rootLogger.error(
+      chalk.red(`No proposer found for transaction ${safeTxHash}`),
+    );
     return;
   }
 
   // Compare proposer to signer
   const signerAddress = await signer.getAddress();
   if (proposer !== signerAddress) {
-    console.log(
+    rootLogger.info(
       chalk.italic(
         `Skipping deletion of transaction ${safeTxHash} proposed by ${proposer}`,
       ),
     );
     return;
   }
-  console.log(`Deleting transaction ${safeTxHash} proposed by ${proposer}`);
+  rootLogger.info(`Deleting transaction ${safeTxHash} proposed by ${proposer}`);
 
   try {
     // Generate the EIP-712 signature
@@ -275,7 +285,7 @@ export async function deleteSafeTx(
     });
 
     if (res.status === 204) {
-      console.log(
+      rootLogger.info(
         chalk.green(
           `Successfully deleted transaction ${safeTxHash} on ${chain}`,
         ),
@@ -284,13 +294,13 @@ export async function deleteSafeTx(
     }
 
     const errorBody = await res.text();
-    console.error(
+    rootLogger.error(
       chalk.red(
         `Failed to delete transaction ${safeTxHash} on ${chain}: Status ${res.status} ${res.statusText}. Response body: ${errorBody}`,
       ),
     );
   } catch (error) {
-    console.error(
+    rootLogger.error(
       chalk.red(`Failed to delete transaction ${safeTxHash} on ${chain}:`),
       error,
     );
@@ -310,8 +320,8 @@ export async function updateSafeOwner(
     (newOwner) => !owners.some((owner) => eqAddress(newOwner, owner)),
   );
 
-  console.log(chalk.magentaBright('Owners to remove:', ownersToRemove));
-  console.log(chalk.magentaBright('Owners to add:', ownersToAdd));
+  rootLogger.info(chalk.magentaBright('Owners to remove:', ownersToRemove));
+  rootLogger.info(chalk.magentaBright('Owners to add:', ownersToAdd));
 
   const transactions: AnnotatedCallData[] = [];
 
@@ -342,4 +352,105 @@ export async function updateSafeOwner(
   }
 
   return transactions;
+}
+
+type SafeStatus = {
+  chain: string;
+  nonce: number;
+  submissionDate: string;
+  shortTxHash: string;
+  fullTxHash: string;
+  confs: number;
+  threshold: number;
+  status: string;
+  balance: string;
+};
+
+export enum SafeTxStatus {
+  NO_CONFIRMATIONS = '🔴',
+  PENDING = '🟡',
+  ONE_AWAY = '🔵',
+  READY_TO_EXECUTE = '🟢',
+}
+
+export async function getPendingTxsForChains(
+  chains: string[],
+  multiProvider: MultiProvider,
+  safes: Record<string, Address>,
+): Promise<SafeStatus[]> {
+  const txs: SafeStatus[] = [];
+  await Promise.all(
+    chains.map(async (chain) => {
+      if (!safes[chain]) {
+        rootLogger.error(chalk.red.bold(`No safe found for ${chain}`));
+        return;
+      }
+
+      if (chain === 'endurance') {
+        rootLogger.info(
+          chalk.gray.italic(
+            `Skipping chain ${chain} as it does not have a functional safe API`,
+          ),
+        );
+        return;
+      }
+
+      let safeSdk, safeService;
+      try {
+        ({ safeSdk, safeService } = await getSafeAndService(
+          chain,
+          multiProvider,
+          safes[chain],
+        ));
+      } catch (error) {
+        rootLogger.warn(
+          chalk.yellow(
+            `Skipping chain ${chain} as there was an error getting the safe service: ${error}`,
+          ),
+        );
+        return;
+      }
+
+      const threshold = await safeSdk.getThreshold();
+      const pendingTxs = await safeService.getPendingTransactions(safes[chain]);
+      if (pendingTxs.results.length === 0) {
+        return;
+      }
+
+      const balance = await safeSdk.getBalance();
+      const nativeToken = await multiProvider.getNativeToken(chain);
+      const formattedBalance = formatUnits(balance, nativeToken.decimals);
+
+      pendingTxs.results.forEach(
+        ({ nonce, submissionDate, safeTxHash, confirmations }) => {
+          const confs = confirmations?.length ?? 0;
+          const status =
+            confs >= threshold
+              ? SafeTxStatus.READY_TO_EXECUTE
+              : confs === 0
+              ? SafeTxStatus.NO_CONFIRMATIONS
+              : threshold - confs
+              ? SafeTxStatus.ONE_AWAY
+              : SafeTxStatus.PENDING;
+
+          txs.push({
+            chain,
+            nonce,
+            submissionDate: new Date(submissionDate).toDateString(),
+            shortTxHash: `${safeTxHash.slice(0, 6)}...${safeTxHash.slice(-4)}`,
+            fullTxHash: safeTxHash,
+            confs,
+            threshold,
+            status,
+            balance: `${Number(formattedBalance).toFixed(5)} ${
+              nativeToken.symbol
+            }`,
+          });
+        },
+      );
+    }),
+  );
+  return txs.sort(
+    (a, b) => a.chain.localeCompare(b.chain) || a.nonce - b.nonce,
+  );
 }
